requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
docopt==0.6.2
elastic-transport==8.17.1
elasticsearch==8.17.1
feedparser==6.0.11
fonttools==4.57.0
h11==0.14.0
hdfs==2.7.3
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
kafka-python==2.0.2
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.4
openai==1.61.1
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
py4j==0.10.9.7
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.3
pyspark==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
